address check issues in tests
