+++ tests/pre_merge/utils/callbacks/metrics_configuration_callback/test_metrics_configuration_callback.py
@@ -1,52 +1,16 @@
-import shutil
-import tempfile
 from pathlib import Path
 
 import pytest
 import pytorch_lightning as pl
-import torch
 from omegaconf import OmegaConf
-from torch import nn
-from torch.utils.data import DataLoader, Dataset
 
 from anomalib.models.components import AnomalyModule
 from anomalib.utils.callbacks.metrics_configuration import MetricsConfigurationCallback
-from anomalib.utils.loggers.tensorboard import AnomalibTensorBoardLogger
 from anomalib.utils.metrics.collection import AnomalibMetricCollection
+from tests.helpers.dummy import DummyDataModule, DummyLogger, DummyModel
 
 
-class DummyDataset(Dataset):
-    def __init__(self):
-        super().__init__()
-
-    def __len__(self):
-        return 1
-
-    def __getitem__(self, idx):
-        return torch.ones(1)
-
-
-class DummyDataModule(pl.LightningDataModule):
-    def test_dataloader(self) -> DataLoader:
-        return DataLoader(DummyDataset())
-
-
-class DummyModel(nn.Module):
-    def __init__(self):
-        super().__init__()
-
-
-class DummyLogger(AnomalibTensorBoardLogger):
-    def __init__(self):
-        self.tempdir = Path(tempfile.mkdtemp())
-        super().__init__(name=f"tensorboard_logs", save_dir=self.tempdir)
-
-    def __del__(self):
-        if self.tempdir.exists():
-            shutil.rmtree(self.tempdir)
-
-
-class DummyAnomalyModule(AnomalyModule):
+class _DummyAnomalyModule(AnomalyModule):
     def __init__(self):
         super().__init__()
         self.model = DummyModel()
@@ -57,7 +21,7 @@ class DummyAnomalyModule(AnomalyModule):
     def test_step(self, batch, _):
         return None
 
-    def validation_epoch_end(self, output):
+    def validation_epoch_end(self, outputs):
         return None
 
     def test_epoch_end(self, outputs):
@@ -85,7 +49,7 @@ def test_metric_collection_configuration_callback(config_from_yaml):
     )
 
     dummy_logger = DummyLogger()
-    dummy_anomaly_module = DummyAnomalyModule()
+    dummy_anomaly_module = _DummyAnomalyModule()
     trainer = pl.Trainer(
         callbacks=[callback], logger=dummy_logger, checkpoint_callback=False, default_root_dir=dummy_logger.tempdir
     )
